refactor(explorer): one structural-row enumeration + shared dnd pointer math
Health-check item 6. The site explorer's drag-and-drop logic was spread
across three verbatim copies:

- structuralRowsForSection (which rows a section contains + their order)
  existed in BOTH useSiteExplorerDnd.ts (current-index lookup) and
  explorerActions.ts (reorder target math) — including copies of
  addFolderPrefixes / parentPathForPath / structuralRowKey / sameParent /
  compareStructuralRows. The two row models could drift, making the panel
  compute a drop index against a different row list than the store reorders.
- getDragPoint / getEventPoint were copied between useDomPanelDnd.ts and
  useSiteExplorerDnd.ts and HAD already drifted (one accepted
  DragMoveEvent | DragEndEvent, the other only DragMoveEvent).

Consolidation:
- @core/page-tree/structuralRows.ts — the one enumeration
  (structuralRowsForSection, structuralRowKey, sameStructuralParent,
  compareStructuralRows, StructuralExplorerRow); both the hook and the
  store action import it. The canonical filter uses isHomePage (the
  store copy hand-rolled page.slug === 'index').
- @core/page-tree/explorerPaths.ts — parentPathForPath/addFolderPrefixes
  shared by structuralRows + the siteExplorer reconciler (whose private
  copies are deleted; also keeps siteExplorer.ts under the 700-line
  module-size ceiling).
- src/admin/lib/dndPointer.ts — getDragPoint/getEventPoint shared by both
  panels, with the wider (correct) event union.

Behavior preserved: 5,359 tests pass incl. siteExplorerPanel.test.tsx and
dom-panel-dnd target-resolution suites; build + lint green.
